UI: Refactor reducers and core controller
Refactor core controller to support multiple redux reducers

Change-type: patch
