+++ lib/ui/components/TailStreamer.tsx
@@ -3,7 +3,9 @@ import { JSONSchema6 } from 'json-schema';
 import * as _ from 'lodash';
 import * as React from 'react';
 import { Card } from '../../Types';
-import { actionCreators, sdk, store } from '../app';
+import { store } from '../core';
+import { sdk } from '../core/sdk';
+import { actionCreators } from '../core/store';
 import { debug } from '../services/helpers';
 import { loadSchema } from '../services/sdk-helpers';
