+++ lib/ui/core/constants.ts
@@ -1,0 +1,1 @@
+export const STORAGE_KEY = 'jellyfish_store';
