+++ lib/ui/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
-import { store } from './app';
+import { store } from './core';
 import { JellyfishUI } from './JellyfishUI';
 
 ReactDOM.render(
